# Chore/M2-7771 improve encrypt and removing wait from use upload progress