### 📝 Description

<!-- Contributions are welcome! If there is a corresponding      -->
<!-- JIRA ticket, link to it by replacing `#` with ticket number -->

🔗 [Jira Ticket M2-7771](https://mindlogger.atlassian.net/browse/M2-7771)
in encrypt function  it was constantly using Buffer.concat.
By modifying  the function to collect encrypted chunks in an array and concatenating them once at the end we can improve the performance.
Removing the wait, we can improve even more the time.

Changes include:

- removing wait from the useUploadProgress
- improving encrypt function



### 🪤 Peer Testing

As long you already have the app properly configured, you must create an applet and test if the save flow is working properly and the time reduced for at least 1 second.


### ✏️ Notes
Before, the upload answers was taking about 1400 to 1600ms and encrypt was taking 1578 to 1650ms