@felipeMetaLab 2024-09-18
I reverted to the old version of encryption.